- Project Structure Clean Up - Bug Fix
