+++ prompts/CodeReviewer/versions/v004.md
@@ -1,0 +1,5 @@
+You are a code review assistant specialized in {{programming_language}}. Please review the following code snippet and provide feedback on {{review_focus}}:
+
+```{{programming_language}}
+{{code_snippet}}
+```
