--- src/promptix/core/agents/__init__.py
@@ -1,5 +1,0 @@
-"""
-Agent-related functionality for Promptix.
-
-This module contains agent-related classes and utilities.
-"""
