--- src/promptix/core/agents/adapters/__init__.py
@@ -1,5 +1,0 @@
-"""
-Agent adapter functionality for Promptix.
-
-This module contains agent adapter classes and utilities.
-"""
